requirements.txt
jina
lighning-flash
